解决文件夹上传大部分问题 1.上传文件夹返回文件夹的id 2.多层级嵌套的子文件附属

+++ src/main/java/uno/acloud/controller/UploadController.java
@@ -43,7 +43,7 @@ public class UploadController {
         Map<String, Object> claims = JwtUtils.parseJWT(jwt);
         int userId = (Integer) claims.get("userId");
 
-        Integer folderId = uploadService.uploadFolder(folderName, parentId, userId);
+        Long folderId = uploadService.uploadFolder(folderName, parentId, userId);
         if (folderId != null) {
             log.info("上传文件夹{}成功", folderName);
             return Result.success(folderId);
